requirements.txt
glfw==2.5.0
numpy==1.21.5
pyopengl==3.1.5
pillow==8.4.0
pandas==1.3.5
mpmath==1.2.1
